requirements_v2.txt
# VOKG v2 Requirements
# Video Object Knowledge Graph with Semantic Understanding

# Core Dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=9.0.0
scipy>=1.7.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# CLIP for Semantic Labeling
git+https://github.com/openai/CLIP.git

# Vector Search
faiss-cpu>=1.7.0  # Use faiss-gpu if CUDA available

# LLM Integration (choose one or both)
openai>=1.0.0              # For GPT-4
google-generativeai>=0.3.0  # For Gemini

# Utilities
tqdm>=4.65.0
pydantic>=2.0.0

# Optional: For production deployment
# celery>=5.3.0
# redis>=4.5.0
# psycopg2-binary>=2.9.0
# neo4j>=5.0.0

# Development
pytest>=7.4.0
black>=23.0.0
